# api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: TrackBot App API
  description: |
    Comprehensive API for the TrackBot Application helping in logging workouts, exercises and nutrition using AI.
    
    This API provides endpoints for:
    - User management
    - Workout tracking and management
    - Exercise logging (cardio and weight training)
    - AI conversation and messaging
    
    All timestamps are in ISO 8601 format (UTC).
  version: 1.0.0
  contact:
    name: Utkarsh Khandelwal
    email: utkarshkhandelwal2011@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.trackbotapp.com/v1
    description: Production server
...
paths:
...
  /workouts/{workoutId}/exercises:
    get:
      summary: List exercises for a workout
      description: Retrieve all exercises for a specific workout
      operationId: listExercises
      tags: [Exercises]
      parameters:
        - $ref: './components/parameters/common.yaml#/WorkoutIdPath'
      responses:
        '200':
          description: List of exercises retrieved successfully
          content:
            application/json:
              schema:
                $ref: './components/schemas/exercise.yaml#/components/schemas/ListExercisesResponse'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
    post:
      summary: Create a new exercise
      description: Add a new exercise to a workout (cardio or weight training)
      operationId: createExercise
      tags: [Exercises]
      parameters:
        - $ref: './components/parameters/common.yaml#/WorkoutIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas/exercise.yaml#/components/schemas/CreateExerciseRequest'
      responses:
        '201':
          description: Exercise created successfully
          content:
            application/json:
              schema:
                $ref: './components/schemas/exercise.yaml#/components/schemas/Exercise'
        '400':
          $ref: './components/responses/errors.yaml#/components/responses/BadRequest'
        '404':
          $ref: './components/responses/errors.yaml#/components/responses/NotFound'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
  /exercises/{exerciseId}:
    get:
      summary: Get exercise by ID
      description: Retrieve a specific exercise by its ID
      operationId: getExerciseById
      tags: [Exercises]
      parameters:
        - $ref: './components/parameters/common.yaml#/ExerciseIdPath'
      responses:
        '200':
          description: Exercise details retrieved successfully
          content:
            application/json:
              schema:
                $ref: './components/schemas/exercise.yaml#/components/schemas/Exercise'
        '404':
          $ref: './components/responses/errors.yaml#/components/responses/NotFound'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
    put:
      summary: Update exercise
      description: Update an existing exercise
      operationId: updateExercise
      tags: [Exercises]
      parameters:
        - $ref: './components/parameters/common.yaml#/ExerciseIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas/exercise.yaml#/components/schemas/CreateExerciseRequest'
      responses:
        '200':
          description: Exercise updated successfully
          content:
            application/json:
              schema:
                $ref: './components/schemas/exercise.yaml#/components/schemas/Exercise'
        '400':
          $ref: './components/responses/errors.yaml#/components/responses/BadRequest'
        '404':
          $ref: './components/responses/errors.yaml#/components/responses/NotFound'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
    delete:
      summary: Delete exercise
      description: Delete an exercise from a workout
      operationId: deleteExercise
      tags: [Exercises]
      parameters:
        - $ref: './components/parameters/common.yaml#/ExerciseIdPath'
      responses:
        '204':
          $ref: './components/responses/errors.yaml#/components/responses/NoContent'
        '404':
          $ref: './components/responses/errors.yaml#/components/responses/NotFound'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'

  # Conversation endpoints
...
components:
  schemas:
...
    Exercise:
      $ref: './components/schemas/exercise.yaml#/components/schemas/Exercise'
...
    CreateExerciseRequest:
      $ref: './components/schemas/exercise.yaml#/components/schemas/CreateExerciseRequest'
    ListExercisesResponse:
      $ref: './components/schemas/exercise.yaml#/components/schemas/ListExercisesResponse'
...
  responses:
    BadRequest:
      $ref: './components/responses/errors.yaml#/components/responses/BadRequest'
...
    NotFound:
      $ref: './components/responses/errors.yaml#/components/responses/NotFound'
...
    InternalServerError:
      $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
    NoContent:
      $ref: './components/responses/errors.yaml#/components/responses/NoContent'